log: filter functions by name
Each log line has an entry of the form at=main.go:123
added to it by Printkv. It's not helpful for this
file-line information to point to a logging helper
function (such as log.Printf or core.logHTTPError), so
we had mechanisms for supplying an alternative:

- Outside of package log, helper functions were expected
to supply an at=foo.go:456 key-value pair to override
the automatically-added one. In practice, none ever did.

- Within package log, the function caller took a "skip"
argument specifying how many frames to skip before
returning file-line information. Helper functions Printf
and Error used this to add override entries.

This mechanism was both brittle and little-used.
Hopefully, this new mechanism will be less brittle and
used where appropriate.

The new approach keeps a lookup table of names of
functions to avoid. Package log then walks up the call
stack, skipping any filtered functions, and prints the
first unfiltered location it finds in the log output.